refactor: FeignClient 클래스 이름 및 패키지 변경
