+++ user-service/src/main/java/com/palja/user_service/application/port/ProductClient.java
@@ -1,0 +1,9 @@
+package com.palja.user_service.application.port;
+
+import java.util.UUID;
+
+public interface ProductClient {
+
+	void deleteAllProducts(UUID companyUserId);
+
+}
