+++ user-service/src/main/java/com/palja/user_service/application/port/ReviewClient.java
@@ -1,0 +1,7 @@
+package com.palja.user_service.application.port;
+
+public interface ReviewClient {
+
+	void deleteAllReviews(Long userId);
+
+}
