--- user-service/src/main/java/com/palja/user_service/application/service/ProductService.java
@@ -1,9 +1,0 @@
-package com.palja.user_service.application.service;
-
-import java.util.UUID;
-
-public interface ProductService {
-
-	void deleteAllProducts(UUID companyUserId);
-
-}
