--- user-service/src/main/java/com/palja/user_service/application/service/ReviewService.java
@@ -1,7 +1,0 @@
-package com.palja.user_service.application.service;
-
-public interface ReviewService {
-
-	void deleteAllReviews(Long userId);
-
-}
