--- user-service/src/main/java/com/palja/user_service/application/service/TimeDealService.java
@@ -1,9 +1,0 @@
-package com.palja.user_service.application.service;
-
-import java.util.UUID;
-
-public interface TimeDealService {
-
-	void deleteAllTimeDeals(UUID companyUserId);
-
-}
