+++ user-service/src/main/java/com/palja/user_service/application/service/impl/CompanyUserServiceImpl.java
@@ -22,9 +22,9 @@ import com.palja.user_service.application.dto.response.ReadCompanyUserSummaryRes
 import com.palja.user_service.application.dto.response.UpdateCompanyUserDetailRes;
 import com.palja.user_service.application.exception.AuthErrorCode;
 import com.palja.user_service.application.exception.UserErrorCode;
+import com.palja.user_service.application.port.ProductClient;
+import com.palja.user_service.application.port.TimeDealClient;
 import com.palja.user_service.application.service.CompanyUserService;
-import com.palja.user_service.application.service.ProductService;
-import com.palja.user_service.application.service.TimeDealService;
 import com.palja.user_service.application.util.JwtUtil;
 import com.palja.user_service.domain.entity.CompanyUser;
 import com.palja.user_service.domain.entity.User;
@@ -43,8 +43,8 @@ public class CompanyUserServiceImpl implements CompanyUserService {
 	private final UserRepository userRepository;
 	private final TokenRepository tokenRepository;
 
-	private final TimeDealService timeDealService;
-	private final ProductService productService;
+	private final TimeDealClient timeDealClient;
+	private final ProductClient productClient;
 
 	private final PasswordEncoder passwordEncoder;
 	private final JwtUtil jwtUtil;
@@ -137,8 +137,8 @@ public class CompanyUserServiceImpl implements CompanyUserService {
 
 		CompanyUser companyUser = getCompanyUserByLoginId(loginId);
 		companyUser.softDelete();
-		timeDealService.deleteAllTimeDeals(companyUser.getId());
-		productService.deleteAllProducts(companyUser.getId());
+		timeDealClient.deleteAllTimeDeals(companyUser.getId());
+		productClient.deleteAllProducts(companyUser.getId());
 	}
 
 	@Override
@@ -148,8 +148,8 @@ public class CompanyUserServiceImpl implements CompanyUserService {
 
 		CompanyUser companyUser = getCompanyUserByLoginId(currentUserLoginId);
 		companyUser.softDelete();
-		timeDealService.deleteAllTimeDeals(companyUser.getId());
-		productService.deleteAllProducts(companyUser.getId());
+		timeDealClient.deleteAllTimeDeals(companyUser.getId());
+		productClient.deleteAllProducts(companyUser.getId());
 
 		String substringAccessToken = jwtUtil.substringToken(accessToken);
 		String hashKey = jwtUtil.hashingTokenToSHA256(substringAccessToken);
