+++ user-service/src/main/java/com/palja/user_service/application/service/impl/CustomerServiceImpl.java
@@ -19,8 +19,8 @@ import com.palja.user_service.application.dto.response.ReadCustomerSummaryRes;
 import com.palja.user_service.application.dto.response.UpdateCustomerDetailRes;
 import com.palja.user_service.application.exception.AuthErrorCode;
 import com.palja.user_service.application.exception.UserErrorCode;
+import com.palja.user_service.application.port.ReviewClient;
 import com.palja.user_service.application.service.CustomerService;
-import com.palja.user_service.application.service.ReviewService;
 import com.palja.user_service.application.util.JwtUtil;
 import com.palja.user_service.domain.entity.User;
 import com.palja.user_service.domain.repository.TokenRepository;
@@ -35,7 +35,7 @@ public class CustomerServiceImpl implements CustomerService {
 	private final UserRepository userRepository;
 	private final TokenRepository tokenRepository;
 
-	private final ReviewService reviewService;
+	private final ReviewClient reviewClient;
 
 	private final PasswordEncoder passwordEncoder;
 	private final JwtUtil jwtUtil;
@@ -117,7 +117,7 @@ public class CustomerServiceImpl implements CustomerService {
 
 		User user = getCustomerByLoginId(loginId);
 		user.softDelete();
-		reviewService.deleteAllReviews(user.getId());
+		reviewClient.deleteAllReviews(user.getId());
 	}
 
 	@Override
@@ -127,7 +127,7 @@ public class CustomerServiceImpl implements CustomerService {
 
 		User user = getCustomerByLoginId(currentUserLoginId);
 		user.softDelete();
-		reviewService.deleteAllReviews(user.getId());
+		reviewClient.deleteAllReviews(user.getId());
 
 		String substringAccessToken = jwtUtil.substringToken(accessToken);
 		String hashKey = jwtUtil.hashingTokenToSHA256(substringAccessToken);
