+++ user-service/src/main/java/com/palja/user_service/infrastructure/external/feign/adapter/ReviewAdapter.java
@@ -5,8 +5,8 @@ import org.springframework.stereotype.Component;
 import com.palja.common.exception.BusinessException;
 import com.palja.common.exception.CommonErrorCode;
 import com.palja.common.response.ApiResponse;
-import com.palja.user_service.application.service.ReviewService;
-import com.palja.user_service.infrastructure.external.feign.ReviewClient;
+import com.palja.user_service.application.port.ReviewClient;
+import com.palja.user_service.infrastructure.external.feign.ReviewFeignClient;
 
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
@@ -15,15 +15,15 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ReviewAdapter implements ReviewService {
+public class ReviewAdapter implements ReviewClient {
 
-	private final ReviewClient reviewClient;
+	private final ReviewFeignClient reviewFeignClient;
 
 	@Override
 	public void deleteAllReviews(Long userId) {
 		try {
 			getDummy(userId); // TODO: API 개발 완료 후 실제 호출로 변경
-			// reviewClient.deleteAllReviews(userId);
+			// reviewFeignClient.deleteAllReviews(userId);
 		} catch (FeignException e) {
 			log.error("[Feign] status={} url=[{}] {} message={}",
 				e.status(), e.request().httpMethod().name(), e.request().url(), e.contentUTF8());
