+++ user-service/src/main/java/com/palja/user_service/infrastructure/external/feign/adapter/TimeDealAdapter.java
@@ -7,8 +7,8 @@ import org.springframework.stereotype.Component;
 import com.palja.common.exception.BusinessException;
 import com.palja.common.exception.CommonErrorCode;
 import com.palja.common.response.ApiResponse;
-import com.palja.user_service.application.service.TimeDealService;
-import com.palja.user_service.infrastructure.external.feign.TimeDealClient;
+import com.palja.user_service.application.port.TimeDealClient;
+import com.palja.user_service.infrastructure.external.feign.TimeDealFeignClient;
 
 import feign.FeignException;
 import lombok.RequiredArgsConstructor;
@@ -17,15 +17,15 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class TimeDealAdapter implements TimeDealService {
+public class TimeDealAdapter implements TimeDealClient {
 
-	private final TimeDealClient timeDealClient;
+	private final TimeDealFeignClient timeDealFeignClient;
 
 	@Override
 	public void deleteAllTimeDeals(UUID companyUserId) {
 		try {
 			getDummy(companyUserId); // TODO: API 개발 완료 후 실제 호출로 변경
-			// timeDealClient.deleteAllTimeDeals(companyUserId);
+			// timeDealFeignClient.deleteAllTimeDeals(companyUserId);
 		} catch (FeignException e) {
 			log.error("[Feign] status={} url=[{}] {} message={}",
 				e.status(), e.request().httpMethod().name(), e.request().url(), e.contentUTF8());
