+++ user-service/src/test/java/com/palja/user_service/service/CompanyUserServiceImplTest.java
@@ -34,8 +34,8 @@ import com.palja.user_service.application.dto.response.CreateUserRes;
 import com.palja.user_service.application.dto.response.ReadCompanyUserDetailRes;
 import com.palja.user_service.application.dto.response.ReadCompanyUserSummaryRes;
 import com.palja.user_service.application.dto.response.UpdateCompanyUserDetailRes;
-import com.palja.user_service.application.service.ProductService;
-import com.palja.user_service.application.service.TimeDealService;
+import com.palja.user_service.application.port.ProductClient;
+import com.palja.user_service.application.port.TimeDealClient;
 import com.palja.user_service.application.service.impl.CompanyUserServiceImpl;
 import com.palja.user_service.application.util.JwtUtil;
 import com.palja.user_service.domain.entity.CompanyUser;
@@ -53,8 +53,8 @@ public class CompanyUserServiceImplTest {
 	@Mock private CompanyUserRepository companyUserRepository;
 	@Mock private UserRepository userRepository;
 	@Mock private TokenRepository tokenRepository;
-	@Mock private TimeDealService timeDealService;
-	@Mock private ProductService productService;
+	@Mock private TimeDealClient timeDealClient;
+	@Mock private ProductClient productClient;
 	@Mock private PasswordEncoder passwordEncoder;
 	@Mock private JwtUtil jwtUtil;
